refactor(dev-app): split probe workspace shell
